+++ inst/examples/dynamicMapLayers.R
@@ -1,9 +1,53 @@
+#' ---
+#' title: "Esri Dynamic Map Layers"
+#' author: "Bhaskar V. Karambelkar"
+#' ---
+
 library(leaflet.esri)
+
+#' ### Example 1
+#' Custom Javascript for identifying features.<br/><br/>
 
 leaflet() %>% setView(-99.88, 37.71, 4) %>%
   addEsriBasemapLayer(esriBasemapLayers$Gray, autoLabels = T) %>%
+  addControl(html = htmltools::HTML(
+    "Click on the map for Soil <a href='https://services.arcgisonline.com/arcgis/rest/services/Specialty/Soil_Survey_Map/MapServer'>Order/ Sub-Order"),
+    layerId = 'selectedFeatures', position = 'bottomleft') %>%
   addEsriDynamicMapLayer(
-    url='https://services.arcgisonline.com/arcgis/rest/services/Specialty/Soil_Survey_Map/MapServer')
+    url='https://services.arcgisonline.com/arcgis/rest/services/Specialty/Soil_Survey_Map/MapServer',
+    layerId = 'soil') %>%
+  htmlwidgets::onRender(htmlwidgets::JS(
+    "function(el, x, data) {
+      var map = this;
+      var identifiedFeature;
+      var pane = document.getElementById('selectedFeatures');
+
+      var soil =  map.layerManager._byLayerId['tile\\nsoil'];
+
+      map.on('click', function (e) {
+        pane.innerHTML = 'Loading';
+        if (identifiedFeature){
+          map.removeLayer(identifiedFeature);
+        }
+        soil.identify().on(map).at(e.latlng).run(function(error, featureCollection){
+          // make sure at least one feature was identified.
+          if (featureCollection.features.length > 0) {
+            identifiedFeature = L.geoJson(featureCollection.features[0]).addTo(map);
+            var soilDescription =
+              featureCollection.features[0].properties['Dominant Order'] +
+              ' - ' +
+              featureCollection.features[0].properties['Dominant Sub-Order'];
+            pane.innerHTML = soilDescription;
+          }
+          else {
+            pane.innerHTML = 'No features identified.';
+          }
+        });
+      });
+    }"))
+
+#' ### Example 2
+#' Custom popup function.<br/><br/>
 
 popupFunc <- htmlwidgets::JS(
   "function (error, featureCollection) {