+++ NEWS.md
@@ -1,11 +1,19 @@
 # leaflet.esri 0.1.3
 
-* Support for DynamicMapLayer and ImageMapLayer
+## Major changes
+
+- Support for DynamicMapLayer
+- Support for ImageMapLayer
 
 # leaflet.esri 0.1.2
 
-* Support for FeatureLayer & TiledMapLayer.
+## Major changes
+
+- Support for FeatureLayer
+- Support for TiledMapLayer
 
 # leaflet.esri 0.1.0
 
-* Initial Release with Esri basemap layers support.
+## Major changes
+
+- Initial Release with Esri basemap layers support